Create summary
The app works to provide an online educational social media platform to aid students with their studies, especially in the times of this pandemic. 
The app would include basic features to form chat groups, share pictures and attach documents as well as features to hold video calls/ online classes. A feature to link a question from the app's own study bank (containing questions sorted chapterwise and difficulty-wise for eg: ranging from cbse level to jee advanced level) would also be included.
The study bank would be one of the main features of the app, thought to contain thousands of questions on every topic. As the student progresses through the exercises, the following recommended would contain more of the topics that the student answered incorrectly/ shared as doubts in their chat groups/ took excessive time in answering etc.
A leaderboard of the students studying in the vicinity could also be made available as a possible feature in the future.
Another major feature of the Study-E app would be its ability to act as a scheduling assistant designed specifically for school/college going students, holding their class/exam schedules as well as study times, deadlines for assignments and custom events scheduled by the user. Providing notifications to the user as reminders.
The app could also be programmed later to accommodate pdf versions of books that the students require.